renamed chitter.rb file as peep.rb file
