+++ lib/peep.rb
